refactor: extract user account controller delegate
